backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
requests==2.31.0
pillow==10.1.0
opencv-python==4.8.1.78
ultralytics==8.0.206
torch==2.1.1
torchvision==0.16.1
easyocr==1.7.6
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
jinja2==3.1.2
email-validator==2.1.0
httpx==0.25.2
websockets==12.0
